--- backend/src/main/java/ubb/courses/backend/controllers/course/ICourseController.java
@@ -1,10 +1,0 @@
-package ubb.courses.backend.controllers.course;
-
-import org.springframework.http.ResponseEntity;
-import ubb.courses.backend.dtos.courses.CourseDTO;
-
-import java.util.Collection;
-
-public interface ICourseController {
-    ResponseEntity<Collection<CourseDTO>> getAllCourses();
-}
